+++ src/sqlx_ext/mod.rs
@@ -1,3 +1,4 @@
+mod sparsevec;
 mod vector;
 
 #[cfg(feature = "halfvec")]